refactor with new modal
