+++ src/css/_settings-modal.scss
@@ -2,9 +2,8 @@
   min-width: 500px;
 
   .settings-form {
-    background: white;
     border: 1px solid $white-3;
-    border-radius: 10px;
+    border-radius: 3px;
 
     a {
       text-decoration: underline;
