+++ src/css/settings/_colors.scss
@@ -77,7 +77,7 @@ $chrome-border: $white-5;
   // colors
   --red: #cd1313;
   --green: #399c81;
-  --green-bright: #4EF567;
+  --green-bright: #4ef567;
   --azure: #2f629c;
   --cello: #324862;
   --havelock: #4b91e2;
