+++ src/js/components/ConnectionModals/StatusLight.tsx
