+++ src/js/components/ModalBox/types.ts
@@ -7,7 +7,6 @@ export type ModalButton = {
   click: Function
   icon?: React.ReactNode
   disabled?: boolean
-  isPrimary?: boolean
 }
 export type ModalButtonTemplate = void | string | ModalButton[]
 export type ModalBoxProps = {
